==> 03 Scripts/01 Explore&Clean US Constructions Data.ipynb <==
   # 01 Import Libraries
    

#import libraries
import pandas as pd
import numpy as np
import os
import matplotlib.pyplot as plt
import seaborn as sns
import scipy

# 02 Import Data

#path
path = r'C:\Users\erhil\OneDrive\Career Foundry Program\Immersion Data Analytics Course\A6 - Adv. Analytics & Dashboard Design\Project A6'

#import data
df = pd.read_csv(os.path.join(path, '02 Data', 'Original', 'US_Constructions_Dec21.csv'), index_col = False)


# 03 Explore Data

pd.set_option('display.max_columns', None)

df.shape

df.head()

df.describe()

df.describe().to_csv(os.path.join(path, '04 Analysis', 'Reports','original_table_1.csv'), index = False)


df['State'].value_counts(dropna = False)

df['Temperature(F)'].value_counts().sort_index().tail(15)

# look at subset of data where temperature is really high
df2 = df.loc[df['Temperature(F)'] == 118]

df2

# 04 Clean the Data
## a. check for mixed variables & change if needed
## b. check for duplicates
## c. check for nulls/blanks
## d. check for unclear data/entry adjustments

# a. check for mixed variables
for col in df.columns.tolist():
    weird = (df[[col]].applymap(type) != df[[col]].iloc[0].apply(type)).any(axis = 1)
    if len (df[weird]) > 0:
        print (col)

#change all mixed variables to object/string
df['City'] = df['City'].astype('str')
df['Timezone'] = df['Timezone'].astype('str')
df['Wind_Direction'] = df['Wind_Direction'].astype('str')
df['Sunrise_Sunset'] = df['Sunrise_Sunset'].astype('str')
df['Street'] = df['Street'].astype('str')
df['Civil_Twilight'] = df['Civil_Twilight'].astype('str')
df['Nautical_Twilight'] = df['Nautical_Twilight'].astype('str')
df['Astronomical_Twilight'] = df['Astronomical_Twilight'].astype('str')
df['Zipcode'] = df['Zipcode'].astype('str')
df['Airport_Code'] = df['Airport_Code'].astype('str')
df['Weather_Condition'] = df['Weather_Condition'].astype('str')
df['Weather_Timestamp'] = df['Weather_Timestamp'].astype('str')

#check to make sure no mixed variable types
for col in df.columns.tolist():
    weird = (df[[col]].applymap(type) != df[[col]].iloc[0].apply(type)).any(axis = 1)
    if len (df[weird]) > 0:
        print (col)

# b. check for duplicates by creating a duplicates sub-dataset
df_dups = df[df.duplicated()]

df.shape

df_dups
# no duplicates indicated

# c. check for nulls/blanks
df.isnull().sum()

#export half-cleaned dataset
#no mixed type data & no duplicates
df.to_pickle(os.path.join(path, '02 Data', 'Prepared', 'df_half_clean.pkl'))
