==> 03 Scripts/02 Clean nulls USConstructions.ipynb <==
# Contents:
## 01 Import Libraries
## 02 Import Data
## 03 Explore Nulls
###  a. Create df2 where Temperature(F) = NaN are omitted
###  b. Create df3 where Distance(mi) = 0 are omitted
###  c. Modify df3 where Precipitation(in) = NaN are imputed with 0 in (median value)
###  d. Create df4 where Visibility = NaN are omitted
###  e. Modify df 4 where Wind_Chill(F) = NaN are replaced with Temperature(F) value
## 04 Export clean dataset as half_clean2

# 01 Import libraries

# import libraries
import pandas as pd
import numpy as np
import os
import matplotlib.pyplot as plt
import seaborn as sns
import scipy

# 02 Import data

#path to project folder
path = r'C:\Users\erhil\OneDrive\Career Foundry Program\Immersion Data Analytics Course\A6 - Adv. Analytics & Dashboard Design\Project A6'

#only need the following columns for analysis
vars_list = ['ID', 'Severity', 'Start_Time', 'End_Time', 'Start_Lat', 'Start_Lng', 'Distance(mi)', 'City', 'County', 'State', 'Zipcode', 'Timezone', 'Airport_Code', 'Weather_Timestamp', 'Temperature(F)', 'Wind_Chill(F)', 'Visibility(mi)', 'Wind_Direction', 'Precipitation(in)', 'Weather_Condition']

#import half-prepared dataset
#still need to work on nulls
df = pd.read_pickle(os.path.join(path, '02 Data', 'Prepared', 'df_half_clean.pkl'))

df1 = df[vars_list]

# 03 Explore nulls

df.shape

df1.shape
#27 columns dropped, as not needed for analysis

df1.describe()

#look at nulls in all columns in original data
df1.isnull().sum()

## a. Create df2 without records where Temperature(F) is NaN

Temp_nomt = pd.notnull(df1['Temperature(F)'])
df2 = df1[Temp_nomt]

pd.set_option('display.max_columns', None)


df2.shape
#115,394 records omitted due to not having a temperature (1.87% of original data)

df2.head()

df2.isnull().sum()

#count unique values in each column
df2.nunique()

pd.options.display.max_rows = None

df2['Weather_Condition'].nunique()

df2['Weather_Condition'].value_counts(dropna = False)

pd.options.display.max_rows = 10

df2

## b. Create df3 without records where Distance(mi) is zero

 #create a df with all values where the distance = 0
df_miss_dist = df2.loc[df2['Distance(mi)'] == 0]


# some of the missing data is due to not having an "end" lat/long, others were just not measured.
# this data consists of 374,171 records and represents approximately 6.06% of the original dataset
df_miss_dist.describe()

#create new dataframe with only valid distances
#cumulatively, will have removed 489,565 records which represents 7.93% of original data
df3 = df2.loc[df2['Distance(mi)'] != 0]

df3.shape

pd.options.display.max_rows = None

df3.isnull().sum()

df3.describe()

df3['Precipitation(in)'].value_counts(dropna = False)

## c. Modify df3 Precipitation(in) by imputing median = 0 for NaNs 

df3['Precipitation(in)'].fillna(0, inplace = True)

df3['Precipitation(in)'].value_counts(dropna = False).sort_index()

df3.isnull().sum()

## d. create df4 with null Visibility(mi) records omitted

df3['Visibility(mi)'].value_counts(dropna = False).sort_index()
#since visibility only has 40,038 null records, I will delete these for my analysis

# create a new df with not-null values for visibility
# cumulatively, I have removed 529,603/6,170,627 records, which is 8.58% of total records
Vis_nomt = pd.notnull(df3['Visibility(mi)'])
df4 = df3[Vis_nomt]

df4.isnull().sum()

## e. Modify df4 for all Wind_Chill(F) values that are null, I will replace with the Temperature(F)

df4['Wind_Chill(F)'].fillna(df['Temperature(F)'],inplace=True)

df4.isnull().sum()

#records removed
w = 374171+115394+40038
print (w)

#number of records removed from original total number of records 
x = 6170627 - w
print (x)

#percent of records omitted from new dataset
y = (w/6170627)*100
print (y)

#double-check number of records
df4.shape

# 04 Export cleaned dataset

#export dataset with 20 columns and 529,603/6,170,627 records removed
#distance = 0 (374,171 records), temperature = NaN (115,394 records), visibility = NaN (40,038 records)
df4.to_csv(os.path.join(path, '02 Data', 'Prepared', 'half_clean2.csv'), index = False)
