==> 03 Scripts/03 Clean nulls and final consistency check data US Constructions.ipynb <==
## Contents
## 01 Import Libraries
## 02 Import Data
## 03 Null Check & Consistency Checks
### a. Wind_Direction column make all capitals/abbreviated
### b. Create df_clean with null Wind_Direction records omitted
### c. Modify df_clean with null Weather_Condition records omitted
### d. Modify df_clean with null City records omitted
### e. Check number of records
### f. look at each column for value_counts to better understand the data
#### - start_time and end_time: made HH:MM:SS and omitted further breakdown of seconds new columns: Start_Time_no_ext and End_Time_no_ext
#### - zipcode: omitted the extension, kept only 5 digits new column: Zipcode_no_ext
#### - temperature/wind_chill values >= 134 degrees are not possible, so were replaced by NaN, then omitted
### g. omit zipcode, start_time, end_time (original, messy) columns
## 04 Export clean.csv
## 05 Export clean.describe table

# 01 Import Libraries

# import libraries
import pandas as pd
import numpy as np
import os
import matplotlib.pyplot as plt
import seaborn as sns
import scipy

# 02 Import Data

#import data
path = r'C:\Users\erhil\OneDrive\Career Foundry Program\Immersion Data Analytics Course\A6 - Adv. Analytics & Dashboard Design\Project A6'

#import data
df = pd.read_csv(os.path.join(path, '02 Data', 'Prepared', 'half_clean2.csv'), index_col = False)

# 03 Null & Consistency Checks

df.shape

df.isnull().sum()

df.head()

## a. Clean up Wind_Direction column
### use all capitals and abbreviations

df['Wind_Direction'].value_counts(dropna = False)
# change: South to S, North to N, East to E, West to W, Variable to VAR, and Calm to CALM

df['Wind_Direction'] = df['Wind_Direction'].str.replace('South', 'S')
df['Wind_Direction'] = df['Wind_Direction'].str.replace('West', 'W')
df['Wind_Direction'] = df['Wind_Direction'].str.replace('North', 'N')
df['Wind_Direction'] = df['Wind_Direction'].str.replace('East', 'E')
df['Wind_Direction'] = df['Wind_Direction'].str.replace('Variable', 'VAR')
df['Wind_Direction'] = df['Wind_Direction'].str.replace('Calm', 'CALM')

df['Wind_Direction'].value_counts(dropna = False)

## b. create df_clean with null Wind_Direction records omitted

# Records with NaNs omitted from dataframe: Wind_Direction(62357), Weather_Condition(TBD), City(TBD)
df_clean = df[df['Wind_Direction'].isnull()==False]

df_clean.shape

df_clean.isnull().sum()

## c. modify df_clean with null Weather_Condition records omitted

# Records with NaNs omitted from dataframe: Wind_Direction(62357)completed, Weather_Condition(13867), City(TBD)
df_clean = df_clean[df_clean['Weather_Condition'].isnull()==False]

df_clean.shape

df_clean.isnull().sum()

## d. modify df_clean with null City records omitted

# Records with NaNs omitted from dataframe: Wind_Direction(62357)completed, Weather_Condition(13867)completed,
# and City(268)
df_clean = df_clean[df_clean['City'].isnull()==False]

# in this script, I have omitted 62357 + 13867 + 268 = 76,492 records
# total eliminated columns = 76,492 + 529,603 = 606,095 out of 6,170,627 (9.8% of total data)
df_clean.isnull().sum()

df_clean.describe()

df_clean.shape

df_clean.nunique()


## e. check number of records

# number of records omitted previously and the number omitted in this script
x = 529603 + (268 + 62357 + 13867)
print (x)

# number of records removed subtracted from original total records
y = 6170627 - x
print (y)

# percent of records intact
z = (y * 100)/6170627
print (z)

## f. look at each column for value_counts to better understand the data
### start_time and end_time: made HH:MM:SS and omitted further breakdown of seconds
### zipcode: omitted the extension, kept only 5 digits
### temperature/wind_chill 
### 

df_clean['Severity'].value_counts(dropna = False)


df_clean['Start_Time'].value_counts(dropna = False).sort_index()
# time needs to be cleaned up to consistent format HH:MM:SS

df_clean['Start_Time_no_ext'] = df_clean['Start_Time'].str[:19]


df_clean.head()

df_clean['End_Time'].value_counts(dropna = False).sort_index()
# time needs to be cleaned up to consistent format HH:MM:SS

df_clean['End_Time_no_ext'] = df_clean['End_Time'].str[:19]


df_clean['End_Time_no_ext'].value_counts(dropna = False).sort_index()

df_clean['Start_Lat'].value_counts(dropna = False).sort_index()

df_clean['Start_Lng'].value_counts(dropna = False).sort_index()

df_clean['Distance(mi)'].value_counts(dropna = False).sort_index()
# .000047 miles is approx. 3 inches; width of TX = 773 miles

df_clean['City'].value_counts(dropna = False).sort_index()

df_clean['County'].value_counts(dropna = False).sort_index()

df_clean['Zipcode'].value_counts(dropna = False).sort_index()
# needs to be cleaned up, omit the extension

df_clean['Zipcode_no_ext'] = df_clean['Zipcode'].str[:5]


df_clean.head()

df_clean['Timezone'].value_counts(dropna = False).sort_index()

df_clean['Airport_Code'].value_counts(dropna = False)

df_clean['Weather_Timestamp'].value_counts(dropna = False)

df_clean['Temperature(F)'].value_counts(dropna = False).sort_index()
#no way temperatures of 196 degrees.. do something here..

df_nuts_temps = df_clean.loc[df_clean['Temperature(F)']>=134]

df_nuts_temps.shape

#find all temperatures 134 degrees and above and replace NaN
df_clean.loc[df_clean['Temperature(F)']>=134,'Temperature(F)'] = np.nan

df_clean = df_clean[df_clean['Temperature(F)'].isnull()==False]

df_clean.shape

df_clean['Wind_Chill(F)'].value_counts(dropna = False).sort_index()

df_clean['Visibility(mi)'].value_counts(dropna = False).sort_index()

df_clean['Wind_Direction'].value_counts(dropna = False).sort_index()

df_clean['Precipitation(in)'].value_counts(dropna = False).sort_index()

df_clean['Weather_Condition'].value_counts(dropna = False).sort_index()

df_clean.shape

## g. omit zipcode, start_time, end_time (original, messy) columns

df_clean2 = df_clean.drop(columns=['Zipcode', 'Start_Time', 'End_Time'])

df_clean2.shape

df_clean2.to_csv(os.path.join(path, '02 Data', 'Prepared', 'clean.csv'), index = False)


df_clean2.describe()

df_clean2.describe().to_csv(os.path.join(path, '04 Analysis', 'Reports','clean2.csv'), index = False)

p = (5564502*100)/6170627
print (p)

90.18% of the original data is still intact.
Hawaii is no longer in the cleaned data, as that state only had 12 records in the original data.
