==> 03 Scripts/04 merger.ipynb <==
## contents

## 01 import libraries & data
## 02 clean airport data
## 03 merge airport & US construction data
## 04 check and export

# 01 import libraries & data

import pandas as pd
import numpy as np
import os
import matplotlib.pyplot as plt
import seaborn as sns
import scipy

#create path to project folder
path = r'C:\Users\erhil\OneDrive\Career Foundry Program\Immersion Data Analytics Course\A6 - Adv. Analytics & Dashboard Design\Project A6'

#import both clean (construction) data and airport data
df_a = pd.read_csv(os.path.join(path, '02 Data', 'Original', 'list-of-airports-in-united-states-of-america-hxl-tags-1.csv'), index_col = False)
df = pd.read_csv(os.path.join(path, '02 Data', 'Prepared', 'clean.csv'), index_col = False)

# 02 clean airport data

#look at airport data
df_a.head()

df_a.shape

#clean this up.. row 0 are comments and not necessary (drop it!)
df_a1=df_a.drop(0)

df_a1.head()

#rename column in airport data from ident to Airport_Code to merge
df_a1.rename(columns={'ident': 'Airport_Code'}, inplace=True)


#amend airport dataframe without extra columns including only these listed
df_a1 = df_a1[['Airport_Code', 'type', 'name', 'latitude_deg', 'longitude_deg']]

# merge airport and clean (US Construction) data

#merge data
df_merged = df.merge(df_a1, how = 'left', on=['Airport_Code'], indicator = True)

#check that column headers & data look ok
df_merged.head()

#check numbers 
#(20 columns from clean Construction, including one joined column) + 4 new columns + _merge column = 25)
#same number of records as in clean Construction :)
df_merged.shape

#410,840 construction data did not have an airport match
df_merged['_merge'].value_counts(dropna = False)

#what percent of data will I be left with, omitting the unmatched data
x = (5153662*100)/6170627
print (x)

#checking that the missing latitude_deg of the airport is the same number as 'left_only' merges
df_merged['latitude_deg'].value_counts(dropna = False)

#executive decision: drop unmatched data since part of my analysis is to look at distance from Construction to Airport
df_merged2 = df_merged[df_merged['latitude_deg'].isnull()==False]

df_merged2.shape

#no need for the _merge column, drop it!
df_merged2=df_merged2.drop(columns=['_merge'])

df_merged2.shape

#check values in dataset, make sure still 49 states (not incl AK & HI, incl DC)
df_merged2.nunique()

#check for null values
df_merged2.isnull().sum()

#not needed for analysis
df_merged2 = df_merged2.drop(columns=('Wind_Chill(F)'))

df_merged2.shape

for col in df_merged2.columns.tolist():
    weird = (df_merged2[[col]].applymap(type) != df_merged2[[col]].iloc[0].apply(type)).any(axis = 1)
    if len (df_merged2[weird]) > 0:
        print (col)

#export data
df_merged2.to_csv(os.path.join(path, '02 Data', 'Prepared', 'merged.csv'), index = False)
